feat: added delete assets and auth for assets-service

+++ internal/assets-service/services/items/items.go
@@ -8,6 +8,7 @@ import (
 	"path/filepath"
 
 	"github.com/FeedTheRealm-org/core-service/config"
+	assets_errors "github.com/FeedTheRealm-org/core-service/internal/assets-service/errors"
 	"github.com/FeedTheRealm-org/core-service/internal/assets-service/models"
 	"github.com/FeedTheRealm-org/core-service/internal/assets-service/repositories/bucket"
 	"github.com/FeedTheRealm-org/core-service/internal/assets-service/repositories/items"
@@ -129,10 +130,14 @@ func (is *itemService) GetAllItems() ([]*models.Item, error) {
 	return is.repository.GetAllItems()
 }
 
-func (is *itemService) DeleteSprite(id uuid.UUID) error {
+func (is *itemService) DeleteItem(id uuid.UUID, userId uuid.UUID) error {
 	sprite, err := is.repository.GetItemById(id)
 	if err != nil {
 		return err
+	}
+
+	if sprite.CreatedBy != userId {
+		return assets_errors.NewUnauthorizedError("you are not the owner of this item")
 	}
 
 	if err := is.repository.DeleteSprite(id); err != nil {
